requirements.txt
click==3.3
## !! Could not determine repository location
file-server==0.1
pyzmq==14.5.0
